+++ libs/engine/src/vivarium/engine/framework/values/interface.py
@@ -69,7 +69,7 @@ class ValuesInterface(Interface):
             a population attribute.
         preferred_combiner
             A strategy for combining the source and the results of any calls
-            to mutators in the pipeline. ``vivarium`` provides the strategies
+            to modifiers in the pipeline. ``vivarium`` provides the strategies
             ``replace_combiner`` (the default) and ``list_combiner``, which
             are importable from ``vivarium.engine.framework.values``. Client code
             may define additional strategies as necessary.
@@ -123,7 +123,7 @@ class ValuesInterface(Interface):
             a population attribute.
         preferred_combiner
             A strategy for combining the source and the results of any calls
-            to mutators in the pipeline. ``vivarium`` provides the strategies
+            to modifiers in the pipeline. ``vivarium`` provides the strategies
             ``replace_combiner`` (the default) and ``list_combiner``, which
             are importable from ``vivarium.engine.framework.values``. Client code
             may define additional strategies as necessary.
@@ -181,7 +181,7 @@ class ValuesInterface(Interface):
             a population attribute.
         preferred_combiner
             A strategy for combining the source and the results of any calls
-            to mutators in the pipeline. ``vivarium`` provides the strategies
+            to modifiers in the pipeline. ``vivarium`` provides the strategies
             ``replace_combiner`` (the default) and ``list_combiner``, which
             are importable from ``vivarium.engine.framework.values``. Client code
             may define additional strategies as necessary.